Struct object_detector_rust::bbox::BBox

pub struct BBox {
    pub x: i32,
    pub y: i32,
    pub width: u32,
    pub height: u32,
}

Struct for representing a bounding box

Fields
x: i32

X coordinate of the top-left corner of the bounding box

y: i32

Y coordinate of the top-left corner of the bounding box

width: u32

Width of the bounding box

height: u32

Height of the bounding box

Implementations

impl BBox

pub fn new(x: i32, y: i32, width: u32, height: u32) -> BBox

Creates a new bounding box with the specified coordinates and dimensions

pub fn left(&self) -> i32

Returns the left coordinate of the bounding box

pub fn right(&self) -> i32

Returns the right coordinate of the bounding box

pub fn top(&self) -> i32

Returns the top coordinate of the bounding box

pub fn bottom(&self) -> i32

Returns the bottom coordinate of the bounding box

pub fn intersect_width(&self, other: &BBox) -> u32

Calculates the width of the intersection between two bounding boxes

Examples

use object_detector_rust::bbox::BBox;

let bbox1 = BBox {
    x: 0,
    y: 0,
    width: 100,
    height: 100,
};
let bbox2 = BBox {
    x: 50,
    y: 0,
    width: 100,
    height: 100,
};
let bbox3 = BBox {
    x: 150,
    y: 0,
    width: 100,
    height: 100,
};
let bbox4 = BBox {
    x: -50,
    y: 0,
    width: 100,
    height: 100,
};

assert_eq!(bbox1.intersect_width(&bbox2), 50);
assert_eq!(bbox1.intersect_width(&bbox3), 0);
assert_eq!(bbox1.intersect_width(&bbox4), 50);
assert_eq!(bbox2.intersect_width(&bbox3), 0);
assert_eq!(bbox2.intersect_width(&bbox4), 0);
assert_eq!(bbox3.intersect_width(&bbox4), 0);

pub fn intersect_height(&self, other: &BBox) -> u32

Calculates the height of the intersection between two bounding boxes

Examples

use object_detector_rust::bbox::BBox;

let bbox1 = BBox {
    x: 0,
    y: 0,
    width: 100,
    height: 100,
};
let bbox2 = BBox {
    x: 0,
    y: 50,
    width: 100,
    height: 100,
};
let bbox3 = BBox {
    x: 0,
    y: 150,
    width: 100,
    height: 100,
};
let bbox4 = BBox {
    x: 0,
    y: -50,
    width: 100,
    height: 100,
};

assert_eq!(bbox1.intersect_height(&bbox2), 50);
assert_eq!(bbox1.intersect_height(&bbox3), 0);
assert_eq!(bbox1.intersect_height(&bbox4), 50);
assert_eq!(bbox2.intersect_height(&bbox3), 0);
assert_eq!(bbox2.intersect_height(&bbox4), 0);
assert_eq!(bbox3.intersect_height(&bbox4), 0);

pub fn intersect_area(&self, other: &BBox) -> u32

Calculates the area of the intersection between two bounding boxes

Examples

use object_detector_rust::bbox::BBox;

let bbox1 = BBox {
    x: 0,
    y: 0,
    width: 100,
    height: 100,
};
let bbox2 = BBox {
    x: 50,
    y: 50,
    width: 100,
    height: 100,
};
let bbox3 = BBox {
    x: 150,
    y: 150,
    width: 100,
    height: 100,
};

assert_eq!(bbox1.intersect_area(&bbox2), 2500);
assert_eq!(bbox1.intersect_area(&bbox3), 0);
assert_eq!(bbox2.intersect_area(&bbox3), 0);

pub fn self_area(&self) -> u32

Calculates the area of the bounding box

Examples

use object_detector_rust::bbox::BBox;

let bbox = BBox {
    x: 0,
    y: 0,
    width: 100,
    height: 100,
};

assert_eq!(bbox.self_area(), 10000);

pub fn union_area(&self, other: &BBox) -> u32

Calculates the area of the union between two bounding boxes

Examples

use object_detector_rust::bbox::BBox;

let bbox1 = BBox {
    x: 0,
    y: 0,
    width: 100,
    height: 100,
};
let bbox2 = BBox {
    x: 50,
    y: 50,
    width: 100,
    height: 100,
};

assert_eq!(bbox1.union_area(&bbox2), 17500);

pub fn overlap(&self, other: &BBox) -> f32

Calculates the overlap between two bounding boxes as the ratio of their intersecting area to their union area

Examples

use object_detector_rust::bbox::BBox;

let bbox1 = BBox {
    x: 0,
    y: 0,
    width: 100,
    height: 100,
};
let bbox2 = BBox {
    x: 50,
    y: 50,
    width: 100,
    height: 100,
};
let bbox3 = BBox {
    x: 150,
    y: 150,
    width: 100,
    height: 100,
};

assert!(bbox1.overlap(&bbox2) > 0.14);
assert_eq!(bbox1.overlap(&bbox3), 0.0);
assert_eq!(bbox2.overlap(&bbox3), 0.0);

pub fn merge(&self, other: &BBox) -> BBox

Merges two bounding boxes into a single bounding box that encloses both

Examples

use object_detector_rust::bbox::BBox;

let bbox1 = BBox {
    x: 0,
    y: 0,
    width: 100,
    height: 100,
};
let bbox2 = BBox {
    x: 50,
    y: 50,
    width: 100,
    height: 100,
};
let bbox3 = BBox {
    x: -50,
    y: -50,
    width: 100,
    height: 100,
};

assert_eq!(bbox1.merge(&bbox2), BBox {
    x: 0,
    y: 0,
    width: 150,
    height: 150,
});
assert_eq!(bbox1.merge(&bbox3), BBox {
    x: -50,
    y: -50,
    width: 150,
    height: 150,
});