Trait object_detector_rust::classifier::Classifier

pub trait Classifier<X, Y>: Trainable<X, Y> + Predictable<X, Y> + Debug + PartialEq { }

The Classifier trait defines the behavior of a machine learning model that is capable of classifying data points into one of several categories. This trait is composed of two other traits: Trainable and Predictable.