Module object_detector_rust::dataset

This module contains the DataSet trait for representing datasets of images and labels for training.

Examples

use object_detector_rust::dataset::DataSet;
use object_detector_rust::dataset::DummyDataSet;

let mut dataset = DummyDataSet::new();
match dataset.load() {
    Ok(_) => {
        let (images, labels) = dataset.get_data();
        // Use the images and labels to train a classifier
    },
    Err(e) => println!("Error loading dataset: {}", e),
}

This example shows how to use the DataSet trait to retrieve images and labels for training a classifier.

Structs

DummyDataSet

Dummy dataset that returns a fixed set of images and labels

FolderDataSet

Struct for representing a dataset loaded from a folder

Traits

DataSet

Trait representing a dataset of images and labels for training