Struct object_detector_rust::dataset::FolderDataSet

pub struct FolderDataSet { /* private fields */ }

Struct for representing a dataset loaded from a folder

Implementations

impl FolderDataSet

pub fn new(
path: &str,
default_width: u32,
default_height: u32,
names: Vec<String>
) -> Self

Creates a new dataset from the specified folder

pub fn load_label_names(path: &str) -> Vec<String>

Loads a list of label names from the specified file path.