Trait object_detector_rust::predictable::Predictable

pub trait Predictable<X, Y> {
    fn predict(&self, x: &ArrayView2<'_, X>) -> Result<Array1<Y>, String>;
}

Trait for objects that can make predictions

Required Methods

fn predict(&self, x: &ArrayView2<'_, X>) -> Result<Array1<Y>, String>

Makes predictions on the provided data using the trained SVMClassifier.

Arguments

x - A 2D array of data to make predictions on.

Returns

A Result containing an error string if prediction fails, or a 1D array of predictions if prediction succeeds.