Module object_detector_rust::utils

Utility functions for working with images

Functions

add_hard_negative_samples

generates hard negative samples, see: Hard Negative Mining

crop_bbox

Crops an image to the dimensions specified in the bounding box

draw_bboxes

Draws bounding boxes on an image.

evaluate_accuracy

Computes the accuracy of a model’s predictions

evaluate_precision

Computes the precision of a model’s predictions

extract_data

Transforms a given vector of extracted features and a vector of labels into a tuple of 2D feature array and 1D label array.

generate_annotations

Generates random annotations for the provided annotated image set.

generate_negative_samples

generates negative samples ///