Function object_detector_rust::utils::add_hard_negative_samples

pub fn add_hard_negative_samples(
    annotated_image_set: &mut dyn AnnotatedImageSet,
    detector: &dyn Detector,
    class: Class,
    max_count: Option<usize>,
    width: u32,
    height: u32
)

generates hard negative samples, see: Hard Negative Mining

Arguments

annotated_image_set - The set of annotated images to add the annotations to.

detector - The hard negative mining detector.

class- The class to assign to the hard negative samples.

max_count - An optional maximum number of annotations to generate for all annotated images.

width - The width of the generated bounding boxes.

height - The height of the generated bounding boxes.