Function object_detector_rust::utils::sliding_window

pub fn sliding_window<I: GenericImageView>(
    image: &I,
    window_size: u32,
    step_size: u32
) -> Vec<ImageWindow<SubImage<&I>>>

Returns a vector of image windows for the given image using a sliding window approach.

The sliding window is a square of size window_size pixels that is moved across the image with a step size of step_size pixels in both the x and y directions from left to right and top to bottom. At each position, the window is added to the vector of image windows returned by the function.

Parameters

image: The image to generate image windows for.

window_size: The size of the sliding window in pixels.

step_size - The number of pixels to move the window in both the x and y directions at each iteration

Example

use object_detector_rust::utils::sliding_window;
use image::DynamicImage;

let image = DynamicImage::new_rgba8(100, 100);
let windows = sliding_window(&image, 50, 25);