Module object_detector_rust::prelude

the object-detector-rust prelude

Re-exports

pub use super::bbox::BBox;

pub use super::classifier::Classifier;

pub use super::dataset::AnnotatedImageSet;

pub use super::dataset::DataSet;

pub use super::detection::Detection;

pub use super::detector::Detector;

pub use super::detector::PersistentDetector;

pub use super::feature::Feature;

pub use super::predictable::Predictable;

pub use super::trainable::Trainable;

pub use super::types::AnnotatedImage;

pub use super::types::Annotation;

pub use super::types::Class;

pub use super::window_generator::WindowGenerator;

Structs

BayesClassifier

The BayesClassifier struct provides a Gaussian Naive Bayes classifier.

BriefFeature

Struct for extracting BRIEF (Binary Robust Independent Elementary Features) from an image.

BriefSVMDetector

Struct representing an object detector using HOG features and an SVM classifier

HOGFeature

Struct for extracting HOG features from images.

HOGSVMDetector

Struct representing an object detector using HOG features and an SVM classifier

MemoryDataSet

Struct for representing a dataset in memory

PyramidWindow

Struct for generating sliding windows over an image

RandomForestClassifier

A struct for training and making predictions with a random forest classifier.

SVMClassifier

The SVMClassifier struct represents an SVM classifier with a trained model. The model is an option type, which means it is initially empty. The new function creates a new SVMClassifier instance with an empty model.

SlidingWindow

Struct for generating sliding windows over an image